Add useMFACancelOnEscape hook for escape key handling in MFA pages
